Added functions to create all tables at once, create enrollment record, create course record, populate tables with sample data, begin cmd interface, and execute function
These are some pretty bold edits.

I firstly made it so that a single function call creates all three tables in the main method. This function is now called create_tables.

Secondly, I created functions that create enrollment and course records in their respective tables. These functions are create_enrollment and create_course.

Thirdly, I created a function that populates the tables with sample data. It's useful to have this as a single function call as I usually have this function commented out because it is difficult to reset the database's records.

Fourthly, I created a function that begins the command line interface. The user is prompted to enter an sid and is prompted to enter function after function until they exit.

Fifthly, I created a function that executes a user's functions, like list, enroll, withdraw, etc. I don't know if this is the best approach because it is kind of messy.

Sixthly, I simplified the main function of the program to make it easier to understand.

Finally, I added comments to make the program easier to understand and removed redundant code in some areas.